remove warning and bzero phys page

+++ kernel/x86/hlt.c
@@ -14,7 +14,7 @@ struct process	*process_hlt_creat(void)
 		return NULL;
 	}
 
-	proc->regs.esp = stack + 128;
+	proc->regs.esp = (u32)(stack + 128);
 	proc->regs.eip = (u32)process_hlt_user;
 	proc->regs.cs = GDT_SEG_KCODE;
 	proc->regs.ss = GDT_SEG_KSTACK;
